pyproject.toml
[project]
name = "sparc.client"
dynamic = ["version"]
description = "NIH SPARC Python Client"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
keywords = ["data science", "datasets"]
authors = [
  {name = "Patryk Orzechowski", email = "patryk@upenn.edu"}
]
maintainers = [
  {name = "Patryk Orzechowski", email = "patryk@upenn.edu"}
]
classifiers=[
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Intended Audience :: Information Technology",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3.10",
  "Topic :: Scientific/Engineering",
  "Topic :: Scientific/Engineering :: Bio-Informatics",
  "Topic :: Scientific/Engineering :: Information Analysis",
  "Topic :: Software Development",
  "Topic :: Software Development :: Libraries",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "osparc~=0.5.0",
  "pennsieve2 >= 0.1.2",
  "cmlibs.zinc >= 4.0.0",
  "cmlibs.utils >= 0.6.0",
  "cmlibs.exporter >= 0.9.0",
  "scaffoldmaker >= 0.10.0",
  "mbfxml2ex >= 0.8.3",
  "python-dotenv >= 0.19.0",
]

[project.optional-dependencies]
test = [
  "pytest ~= 7.2",
  "pytest-cov",
  "pytest-mock",
  "flake8",
  "black",
  "responses",
]

[project.urls]
homepage = "https://sparc.science"
documentation = "https://docs.sparc.science/docs/sparc-python-client"
repository = "https://github.com/nih-sparc/sparc.client"
changelog = "https://github.com/nih-sparc/sparc.client/blob/main/CHANGELOG.md"

[tool.black]
line-length = 99
target-version = ['py310']
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | docs
)/
'''

[tool.setuptools.packages.find]
where = ["src", "config"]

[tool.setuptools.package-data]
"*" = ["*.ini", "*.txt"]

[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8.0", "pennsieve2"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
version_file = "src/sparc/client/_version.py"

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = [
  "tests",
]
pythonpath = [
  ".", "src",
]
